refactor: split utils into separate modules
