+++ src/language_tutor/exercise.py
@@ -1,30 +1,27 @@
-"""Utility functions for the Language Tutor app."""
+"""Exercise-related utilities for Language Tutor."""
 
 import re
 import random
-import asyncio
-import nest_asyncio
 from language_tutor.llm import llm
-from PyQt5.QtWidgets import QApplication
+from language_tutor.config import OR_MODEL_NAME, OR_MODEL_NAME_CHECK
+
 
 async def generate_exercise(language, level, exercise_type, definitions):
     """Generate a new language exercise using LiteLLM.
-    
+
     Args:
         language (str): The language code (e.g., "pl" for Polish)
         level (str): The proficiency level (e.g., "A1")
         exercise_type (str): The type of exercise to generate
         definitions (dict): Dictionary containing exercise definitions
-        
+
     Returns:
         tuple: (exercise_text, hints, cost)
     """
-    from language_tutor.config import OR_MODEL_NAME
-    
     # Construct prompt asking for specific formatting
     prompt = f"""Create a short '{exercise_type}' writing exercise for a learner of {language} for a proficiency level {level}.
     The expected length of the writing should be between {definitions[exercise_type]['expected_length'][0]} and {definitions[exercise_type]['expected_length'][1]} words.
-    Random number is {random.randint(1, 10000)} (don't use it, it is just to make the prompt different). 
+    Random number is {random.randint(1, 10000)} (don't use it, it is just to make the prompt different).
 Provide the exercise text and optionally some hints. The requirements for the exercise are:
 '{definitions[exercise_type]['requirements']}'
 You should generate exactly one exercise. It should be a task, not the text of the exercise itself.
@@ -35,8 +32,7 @@ Format the output EXACTLY like this, using these specific headings:
 [The exercise text goes here]
 
 **Hints:**
-[Optional hints go here. You can add useful phrases in addition to the hints. If no hints, write "None."]
-"""
+[Optional hints go here. You can add useful phrases in addition to the hints. If no hints, write "None."]"""
     messages = [{"role": "user", "content": prompt}]
 
     # Make the async API call
@@ -67,7 +63,7 @@ Format the output EXACTLY like this, using these specific headings:
 
 async def check_writing(language, level, exercise_text, writing_input, exercise_type, definitions):
     """Check the user's writing using LiteLLM.
-    
+
     Args:
         language (str): The language code (e.g., "pl" for Polish)
         level (str): The proficiency level (e.g., "A1")
@@ -75,12 +71,10 @@ async def check_writing(language, level, exercise_text, writing_input, exercise_
         writing_input (str): The user's written response
         exercise_type (str): The type of exercise
         definitions (dict): Dictionary containing exercise definitions
-        
+
     Returns:
         tuple: (mistakes, style_errors, recommendations, cost)
     """
-    from language_tutor.config import OR_MODEL_NAME_CHECK
-    
     # Construct prompt for checking, asking for specific format
     prompt = f"""A student learning {language} was given the exercise:
 '{exercise_text}' with the following requirements:
@@ -103,8 +97,7 @@ Format the output EXACTLY like this, using these specific headings:
 [List of stylistic errors, or "None." if none found]
 
 **Recommendations:**
-[List of recommendations, or "None." if none found]
-"""
+[List of recommendations, or "None." if none found]"""
     messages = [{"role": "user", "content": prompt}]
     model_name = OR_MODEL_NAME_CHECK
 
@@ -120,77 +113,10 @@ Format the output EXACTLY like this, using these specific headings:
     mistakes = mistakes_match.group(1).strip() if mistakes_match else "Could not parse."
     style_errors = style_match.group(1).strip() if style_match else "Could not parse."
     recommendations = recs_match.group(1).strip() if recs_match else "Could not parse."
-    
+
     # Return empty strings instead of "None." to avoid displaying it
     mistakes = "" if mistakes.lower() == "none." else mistakes
     style_errors = "" if style_errors.lower() == "none." else style_errors
     recommendations = "" if recommendations.lower() == "none." else recommendations
 
     return mistakes, style_errors, recommendations, cost
-
-
-async def answer_question(model, question, context):
-    """Answer a question using the specified AI model.
-    
-    Args:
-        model (str): The AI model identifier
-        question (str): The user's question
-        context (dict): Dictionary containing context information
-        
-    Returns:
-        tuple: (answer_text, cost)
-    """
-    # Construct the prompt
-    prompt = f"""You are a helpful language learning assistant. The user is learning {context['language']} 
-at {context['level']} level. They are working on a {context['exercise_type']} exercise:
-
-"{context['exercise']}"
-
-The user's question is:
-{question}
-
-Please provide a helpful, educational response focused on language learning."""
-    
-    # Make the API call
-    messages = [{"role": "user", "content": prompt}]
-    response, cost = await llm.completion(model=model, messages=messages)
-    
-    # Get the response
-    answer = response.choices[0].message.content
-    return answer, cost
-
-def run_async(coro, in_q_application=True):
-    """Run an async coroutine from a synchronous method without blocking UI.
-    
-    Args:
-        coro: The coroutine to run
-        in_q_application: Whether running in Qt application
-    """
-    
-    # Apply nest_asyncio to allow nested event loops
-    try:
-        nest_asyncio.apply()
-    except RuntimeError:
-        # If already applied or not needed, continue
-        pass
-    
-    # Get or create an event loop
-    try:
-        loop = asyncio.get_event_loop()
-    except RuntimeError:
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
-    
-    # Create a future for the coroutine
-    future = asyncio.ensure_future(coro)
-    
-    # Process Qt events while waiting for the coroutine to complete
-    if in_q_application:
-        while not future.done():
-            QApplication.processEvents()
-            loop.run_until_complete(asyncio.sleep(0.01))  # Short sleep to avoid CPU hogging
-    else:
-        # If not in Qt application, just run the event loop until the coroutine is done
-        loop.run_until_complete(future)    
-    # Return the result
-    return future.result()
